Move `waitForever` to `shared/utils`
